+++ src/app/(dashboard)/layout.tsx
@@ -1,11 +1,8 @@
 import { redirect } from 'next/navigation';
 
-import { AppSidebar } from '@/components/dashboard/app-sidebar';
-import { SidebarInset, SidebarProvider } from '@/components/ui/sidebar';
-import MainHeader from '@/components/dashboard/main-header';
-
 import { getCurrentDashboardUser } from './current-user';
 import { DashboardUserProvider } from './user-provider';
+import DashboardWrapper from './wrapper';
 
 export default async function DashboardLayout({ children }: { children: React.ReactNode }) {
   const dashboardUser = await getCurrentDashboardUser();
@@ -16,13 +13,7 @@ export default async function DashboardLayout({ children }: { children: React.Re
 
   return (
     <DashboardUserProvider user={dashboardUser.user} role={dashboardUser.role}>
-      <SidebarProvider>
-        <AppSidebar />
-        <SidebarInset>
-          <MainHeader />
-          <div className="p-5 sm:p-7.5 md:p-10">{children}</div>
-        </SidebarInset>
-      </SidebarProvider>
+      <DashboardWrapper>{children}</DashboardWrapper>
     </DashboardUserProvider>
   );
 }