+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - app sidebar shop route links to not include shop within url
 - nav menu group link highlights current page route
 - nav menu group link automatically opens current route sub menu group
+- split dashboard layout into wrapper component for client component to fix hydration error
 
 ### Fixed